pyproject.toml
[project]
name = "screener"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi>=0.111.0",
    "polygon-api-client>=1.13.7",
    "pydantic-settings>=2.2.1",
    "setuptools>=70.0.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/screener"]
